% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BinToDec}
\alias{BinToDec}
\title{Simple function to convert binary string to decimal}
\usage{
BinToDec(x)
}
\arguments{
\item{x}{Input binary string}
}
\value{
number Decimal value of the given binary string
}
\description{
Simple function to convert binary string to decimal
}
\examples{
number <- BinToDec("1011")
}

