% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker.R
\name{Bootstrap_LOOCV_LR_AUC}
\alias{Bootstrap_LOOCV_LR_AUC}
\title{Do bootstrap and LOOCV}
\usage{
Bootstrap_LOOCV_LR_AUC(df, targetVec, nboot = 50)
}
\arguments{
\item{df}{Row is sample, column is feature. Required}

\item{targetVec}{y vector. Required}

\item{nboot}{number of BOOTSTRAP}
}
\value{
bootstrap loocv result dataframe
}
\description{
Do bootstrap and LOOCV
}
\examples{
data('iris')
Bootstrap_LOOCV_LR_AUC(iris[,1:4],
c(rep(1,100), rep(0,50)), nboot = 3)
}
