% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{coreOTU}
\alias{coreOTU}
\title{Compute Core OTUs for the given data matrix}
\usage{
coreOTU(zcounts, otuthreshold = 0.05, prevalence = 0.4)
}
\arguments{
\item{zcounts}{Standardized counts}

\item{otuthreshold}{Abundance cutoff threshold for the OTU to be picked}

\item{prevalence}{Prevalence of the OTU at threshold cutoff among samples}
}
\value{
list containing core OTUs
}
\description{
Compute Core OTUs for the given data matrix
}
\examples{
example_data_dir <- system.file("example/data", package = "PathoStat")
pathoreport_file_suffix <- "-sam-report.tsv"
datlist <- readPathoscopeData(example_data_dir, pathoreport_file_suffix)
countdat <- datlist$countdata
coreotus <- coreOTU(countdat)
}

