% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreOTUModule.R
\name{coreOTUModule}
\alias{coreOTUModule}
\title{Server function for Core OTU Module}
\usage{
coreOTUModule(input, output, session)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{This is the session}
}
\value{
None
}
\description{
Server function for Core OTU Module
}

