% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreOTUModule.R
\name{coreOTUModuleUI}
\alias{coreOTUModuleUI}
\title{UI function for Core OTU Module}
\usage{
coreOTUModuleUI(id, label = "Core OTUs")
}
\arguments{
\item{id}{Namespace for module}

\item{label}{Tab label}
}
\value{
None
}
\description{
UI function for Core OTU Module
}

