% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{coreOTUNormalize}
\alias{coreOTUNormalize}
\title{Compute Empirical Bayes OTU Normalized data}
\usage{
coreOTUNormalize(zcounts, wt = 0.25, otuthreshold = 0.05,
  prevalence = 0.4)
}
\arguments{
\item{zcounts}{counts data to be normalized}

\item{wt}{Weight parameter indicating how much information to borrow
across samples using Empirical Bayes}

\item{prevalence}{Prevalence of the OTU at threshold cutoff among samples}

\item{threshold}{Abundance cutoff threshold for the OTU to be picked}
}
\value{
list containing Empirical Bayes coreOTU Normalized data
}
\description{
Compute Empirical Bayes OTU Normalized data
}
\examples{
example_data_dir <- system.file("example/data", package = "PathoStat")
pathoreport_file_suffix <- "-sam-report.tsv"
datlist <- readPathoscopeData(example_data_dir, pathoreport_file_suffix)
countdat <- datlist$countdata
coreotunormdat <- coreOTUNormalize(countdat)
}

