% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathoStat.R
\name{createPathoStat}
\alias{createPathoStat}
\title{Generates a PathoStat object from the PathoScope reports for further 
analysis using the interactive shiny app}
\usage{
createPathoStat(input_dir = ".", sample_data_file = "sample_data.tsv",
  pathoreport_file_suffix = "-sam-report.tsv")
}
\arguments{
\item{input_dir}{Directory where the tsv files from PathoScope are located}

\item{batch}{Batch covariate}

\item{condition}{Covariates or conditions of interest besides batch}
}
\value{
outputfile The output file with all the statistical plots
}
\description{
Generates a PathoStat object from the PathoScope reports for further 
analysis using the interactive shiny app
}
\examples{
example_data_dir <- system.file("example/data", package = "PathoStat")
pstat <- createPathoStat(input_dir=example_data_dir, 
    sample_data_file="sample_data.tsv")
}

