% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityModule.R
\name{diversityModule}
\alias{diversityModule}
\title{Server function for Diversity Module}
\usage{
diversityModule(input, output, session, pstat)
}
\arguments{
\item{input}{Shiny server input object}

\item{output}{Shiny server output object}

\item{session}{This is the session}

\item{pstat}{PathoStat object}
}
\value{
None
}
\description{
Server function for Diversity Module
}

