% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityModule.R
\name{diversityModuleUI}
\alias{diversityModuleUI}
\title{UI function for Diversity Module}
\usage{
diversityModuleUI(id, label = "Diversity2")
}
\arguments{
\item{id}{Namespace for module}

\item{label}{Tab label}
}
\value{
None
}
\description{
UI function for Diversity Module
}

