% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{findTaxonLevel}
\alias{findTaxonLevel}
\title{Find the taxonomy for the given taxon id}
\usage{
findTaxonLevel(tid)
}
\arguments{
\item{tid}{Given taxon id}
}
\value{
taxonomy LineageEx
}
\description{
Find the taxonomy for the given taxon id
}

