% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{findTaxonLevelData}
\alias{findTaxonLevelData}
\title{Find the taxonomy level data for the given taxon level}
\usage{
findTaxonLevelData(data, taxonLevels, level)
}
\arguments{
\item{data}{Given Relative abundance data}

\item{taxonLevels}{Taxon Levels of all tids}

\item{level}{Selected Taxonomy Level}
}
\value{
taxdata Taxonomy Level Data
}
\description{
Find the taxonomy level data for the given taxon level
}
\keyword{internal}

