% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{findTaxonMat}
\alias{findTaxonMat}
\title{Find the Taxonomy Information Matrix}
\usage{
findTaxonMat(names, taxonLevels)
}
\arguments{
\item{names}{Row names of the taxonomy matrix}

\item{taxonLevels}{Taxon Levels of all tids}
}
\value{
taxmat Taxonomy Information Matrix
}
\description{
Find the Taxonomy Information Matrix
}

