% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{findTaxonomy}
\alias{findTaxonomy}
\title{Find the taxonomy for each taxon ids}
\usage{
findTaxonomy(tids)
}
\arguments{
\item{tids}{Given taxonomy ids}
}
\value{
taxondata Data with the taxonomy information
}
\description{
Find the taxonomy for each taxon ids
}

