% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatTaxTable}
\alias{formatTaxTable}
\title{Format taxonomy table for rendering}
\usage{
formatTaxTable(ttable)
}
\arguments{
\item{ttable}{Taxonomy table}
}
\value{
Formatted table suitable for rendering with. DT::renderDataTable
}
\description{
Format taxonomy table for rendering
}

