% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreOTUModule.R
\name{get_core}
\alias{get_core}
\title{Select rows of OTU matrix that meet given detection and prevalence 
thresholds}
\usage{
get_core(pstat, detection, prevalence)
}
\arguments{
\item{pstat}{PathoStat object}

\item{detection}{An integer specifying the minimum value considered to be 
"detected"}

\item{prevalence}{An integer specifying the minimum number of samples that
must be detected}
}
\value{
Subsetted PathoStat object
}
\description{
Select rows of OTU matrix that meet given detection and prevalence 
thresholds
}

