% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreOTUModule.R
\name{get_coremat}
\alias{get_coremat}
\title{Create core OTU matrix containing number of OTUs detected at varying 
detection and prevalence thresholds.}
\usage{
get_coremat(pstat)
}
\arguments{
\item{pstat}{PathoStat object}
}
\value{
Data frame containing number of OTUs at varying detection and 
prevalence thresholds, with rows corresponding to number of samples and
columns corresponding to detection thresholds. An additional column called
"prev"contains the sample threshold for each row.
}
\description{
Create core OTU matrix containing number of OTUs detected at varying 
detection and prevalence thresholds.
}

