% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreOTUModule.R
\name{get_coremat_lineplot}
\alias{get_coremat_lineplot}
\title{Create line plot from core OTU matrix}
\usage{
get_coremat_lineplot(coremat)
}
\arguments{
\item{coremat}{Core OTU matrix (data.frame)}
}
\value{
Line plot with number of OTUs on the x-axis and detection threshold
on the y-axis. Lines connect data points with the same number of samples.
}
\description{
Create line plot from core OTU matrix
}

