% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAnalysis.R
\name{GET_PAM}
\alias{GET_PAM}
\title{transform cpm counts to presence-absence matrix}
\usage{
GET_PAM(df)
}
\arguments{
\item{df}{Input data object that contains the data to be tested. Required}
}
\value{
df.output object
}
\description{
transform cpm counts to presence-absence matrix
}
\examples{
GET_PAM(df)
}
