% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadPstat}
\alias{loadPstat}
\title{Load the R data(.rda) file with pathostat object}
\usage{
loadPstat(indir = ".", infileName = "pstat_data.rda")
}
\arguments{
\item{indir}{Input Directory of the .rda file}

\item{infileName}{File name of the .rda file}
}
\value{
pstat pathostat object (NULL if it does not exist)
}
\description{
Load the R data(.rda) file with pathostat object
}
\examples{
data_dir <- system.file("data", package = "PathoStat")
infileName <- "pstat_data.rda"
pstat <- loadPstat(data_dir, infileName)
}

