% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log2CPM}
\alias{log2CPM}
\title{Compute log2(counts per mil reads) and library size for each sample}
\usage{
log2CPM(qcounts, lib.size = NULL)
}
\arguments{
\item{qcounts}{quantile normalized counts}

\item{lib.size}{default is colsums(qcounts)}
}
\value{
list containing log2(quantile counts per mil reads) and library sizes
}
\description{
Compute log2(counts per mil reads) and library size for each sample
}
\examples{
example_data_dir <- system.file("example/data", package = "PathoStat")
pathoreport_file_suffix <- "-sam-report.tsv"
datlist <- readPathoscopeData(example_data_dir, pathoreport_file_suffix)
countdat <- datlist$countdata
lcpm <- log2CPM(countdat)
}

