% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker.R
\name{LOOAUC_simple_multiple_noplot_one_df}
\alias{LOOAUC_simple_multiple_noplot_one_df}
\title{LOOCV}
\usage{
LOOAUC_simple_multiple_noplot_one_df(df, targetVec)
}
\arguments{
\item{df}{Row is sample, column is feature. Required}

\item{target.vec}{y vector. Required}
}
\description{
LOOCV
}
\examples{
LOOAUC_simple_multiple_noplot_one_df(df, target.vec)
}
