% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker.R
\name{LOOAUC_simple_multiple_one_df}
\alias{LOOAUC_simple_multiple_one_df}
\title{LOOCV with ROC curve}
\usage{
LOOAUC_simple_multiple_one_df(df, targetVec)
}
\arguments{
\item{df}{Row is sample, column is feature. Required}

\item{targetVec}{y vector. Required}
}
\value{
the ROC
}
\description{
LOOCV with ROC curve
}
\examples{
data('iris')
LOOAUC_simple_multiple_one_df(iris[,1:4],
c(rep(1,100), rep(0,50)))
}
