% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathoStat.R
\name{pathoStat}
\alias{pathoStat}
\title{Statistical Analysis of the PathoScope reports and generates a html report 
and produces interactive shiny app plots}
\usage{
pathoStat(input_dir = ".", batch, condition = NULL,
  report_file = "PathoStat_report.html", report_dir = ".",
  report_option_binary = "111111111", view_report = TRUE,
  interactive = TRUE)
}
\arguments{
\item{input_dir}{Directory where the tsv files from PathoScope are located}

\item{batch}{Batch covariate}

\item{condition}{Covariates or conditions of interest besides batch}

\item{report_file}{Output report file name}

\item{report_dir}{Output report directory path}

\item{report_option_binary}{9 bits Binary String representing the plots to 
display and hide in the report}

\item{view_report}{when TRUE, opens the report in a browser}
}
\value{
outputfile The output file with all the statistical plots
}
\description{
Statistical Analysis of the PathoScope reports and generates a html report 
and produces interactive shiny app plots
}
\examples{
nbatch <- 10
nperbatch <- 10
batch <- rep(1:nbatch, each=nperbatch)
PathoStat(input_dir='.', batch)
}

