% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confRegion.R
\name{plotConfRegion}
\alias{plotConfRegion}
\title{Compute the confidence region for the given proportions}
\usage{
plotConfRegion(p1, p2, size = 100, uselogit = TRUE, n = 10000,
    seed = 1000, jit = FALSE)
}
\arguments{
\item{p1}{Read counts for first taxon}

\item{p2}{Read counts for second taxon}

\item{size}{Total read counts in the sample}

\item{uselogit}{Use logit transformation to compute confidence region}

\item{n}{Total number of simulation points to generate}

\item{seed}{Seed to use in random simulation}

\item{jit}{jitter option (FALSe by default) for the plot}
}
\value{
Confidence region plot
}
\description{
Compute the confidence region for the given proportions
}
\examples{
p1 <- 20
p2 <- 25
size <- 200
plotConfRegion(p1, p2, size, uselogit=FALSE)
}

