% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotHeatmapColor}
\alias{plotHeatmapColor}
\title{Plot heatmap with color bar}
\usage{
plotHeatmapColor(df.input, condition.vec, do.scale = TRUE,
  clusterRow = TRUE, clusterCol = TRUE, displayRowLabels = FALSE,
  displayColumnLabels = TRUE, displayRowDendrograms = TRUE,
  displayColumnDendrograms = TRUE, annotationColors = "auto",
  columnTitle = "Title")
}
\arguments{
\item{df.input}{Input data object that contains the data to be plotted. Required}

\item{condition.vec}{The condition used for plotting the heatmap. Required}

\item{clusterRow}{Cluster the rows. The default is TRUE}

\item{clusterCol}{Cluster the columns. The default is TRUE}

\item{displayRowLabels}{Display the row labels on the heatmap. The default
is TRUE.}

\item{displayColumnLabels}{Display the column labels on the heatmap. The
default is TRUE}

\item{displayRowDendrograms}{Display the row dendrograms on the heatmap. The
default is TRUE}

\item{displayColumnDendrograms}{Display the column dendrograms on the
heatmap. The default is TRUE.}

\item{annotationColors}{Set of annotation colors for color bar. If null,
no color bar is shown. If "auto", then colors will be
added automatically. The default is "auto".}

\item{columnTitle}{Title to be displayed at top of heatmap.}
}
\value{
ComplexHeatmap object
}
\description{
Plot heatmap with color bar
}
\examples{
plotHeatmapColor(df.input, condition.vec,
clusterRow=TRUE, clusterCol=TRUE, displayRowLabels=FALSE,
displayColumnLabels=TRUE, displayRowDendrograms=TRUE,
displayColumnDendrograms=TRUE, annotationColors = "auto",
columnTitle="Title")
}
