% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathoStat.R
\docType{data}
\name{pstat_data}
\alias{pstat}
\alias{pstat_data}
\title{pathostat object generated from example pathoscope report files}
\format{pathostat object extension of phyloseq-class experiment-level object:
\describe{
    \item{otu_table}{OTU table with  41 taxa and 33 samples}
    \item{sample_data}{Sample Data with 33 samples by 18 sample variables}
    \item{tax_table}{Taxonomy Table with 41 taxa by 9 taxonomic ranks}
    \item{sample_data}{Phylogenetic Tree with 41 tips and 40 internal nodes}
}}
\usage{
pstat
}
\value{
pathostat object
}
\description{
This example data consists of 33 samples from a diet study with 11 subjects 
taking 3 different diets in random order
}
\keyword{datasets}

