% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{savePstat}
\alias{savePstat}
\title{Save the pathostat object to R data(.rda) file}
\usage{
savePstat(pstat, outdir = ".", outfileName = "pstat_data.rda")
}
\arguments{
\item{pstat}{pathostat object}

\item{outdir}{Output Directory of the .rda file}

\item{outfileName}{File name of the .rda file}
}
\value{
outfile .rda file
}
\description{
Save the pathostat object to R data(.rda) file
}
\examples{
data_dir_test <- system.file("data", package = "PathoStat")
pstat_test <- loadPstat(indir=data_dir_test,
infileName="pstat_data_2_L1.rda")
outfile <- savePstat(pstat_test)
}
