% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{sizeNormalize}
\alias{sizeNormalize}
\title{Normalize the given data based on library size}
\usage{
sizeNormalize(zcounts)
}
\arguments{
\item{zcounts}{Input counts data matrix}
}
\value{
acounts Normalized counts data matrix
}
\description{
Normalize the given data based on library size
}
\examples{
example_data_dir <- system.file("example/data", package = "PathoStat")
pathoreport_file_suffix <- "-sam-report.tsv"
datlist <- readPathoscopeData(example_data_dir, pathoreport_file_suffix)
countdat <- datlist$countdata
acounts <- sizeNormalize(countdat)
}

