% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarizeTable}
\alias{summarizeTable}
\title{Summarize sample}
\usage{
summarizeTable(pstat)
}
\arguments{
\item{pstat}{Input pstat}
}
\value{
A data.frame object of summary metrics.
}
\description{
Creates a table of summary metrics
}
\examples{
data_dir_test <- system.file("data", package = "PathoStat")
pstat_test <- loadPstat(indir=data_dir_test,
infileName="pstat_data_2_L1.rda")
st.tmp <- summarizeTable(pstat_test)
}
