% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker.R
\name{Bootstrap_LOOCV_LR_AUC}
\alias{Bootstrap_LOOCV_LR_AUC}
\title{Do bootstrap and LOOCV}
\usage{
Bootstrap_LOOCV_LR_AUC(df, target.vec, nboot = 50)
}
\arguments{
\item{df}{Row is sample, column is feature. Required}

\item{target.vec}{y vector. Required}

\item{nboot}{number of BOOTSTRAP}
}
\description{
Do bootstrap and LOOCV
}
\examples{
Bootstrap_LOOCV_LR_AUC(df, target.vec, nboot)
}
