% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathoStat.R
\name{createPathoStat}
\alias{createPathoStat}
\title{Generates a PathoStat object from the PathoScope reports for further 
analysis using the interactive shiny app}
\usage{
createPathoStat(input_dir = ".", sample_data_file = "sample_data.tsv",
    pathoreport_file_suffix = "-sam-report.tsv")
}
\arguments{
\item{input_dir}{Directory where the tsv files from PathoScope are located}

\item{sample_data_file}{Sample Data file with information about samples}

\item{pathoreport_file_suffix}{PathoScope report files suffix}
}
\value{
pstat The pathostat object generated from the given tsv files
}
\description{
Generates a PathoStat object from the PathoScope reports for further 
analysis using the interactive shiny app
}
\examples{
example_data_dir <- system.file("example/data", package = "PathoStat")
pstat <- createPathoStat(input_dir=example_data_dir, 
    sample_data_file="sample_data.tsv")
}

