% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{findTaxonLevel}
\alias{findTaxonLevel}
\title{Find the taxonomy for the given taxon id}
\usage{
findTaxonLevel(tid)
}
\arguments{
\item{tid}{Given taxon id}
}
\value{
taxonomy LineageEx
}
\description{
Find the taxonomy for the given taxon id
}
\examples{
example_data_dir <- system.file("example/data", package = "PathoStat")
pathoreport_file_suffix <- "-sam-report.tsv"
datlist <- readPathoscopeData(example_data_dir, pathoreport_file_suffix)
dat <- datlist$data
ids <- rownames(dat)
tids <- unlist(lapply(ids, FUN = grepTid))
taxonLevel <- findTaxonomy(tids[1])
}

