% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{findTaxonMat}
\alias{findTaxonMat}
\title{Find the Taxonomy Information Matrix}
\usage{
findTaxonMat(names, taxonLevels)
}
\arguments{
\item{names}{Row names of the taxonomy matrix}

\item{taxonLevels}{Taxon Levels of all tids}
}
\value{
taxmat Taxonomy Information Matrix
}
\description{
Find the Taxonomy Information Matrix
}
\examples{
example_data_dir <- system.file("example/data", package = "PathoStat")
pathoreport_file_suffix <- "-sam-report.tsv"
datlist <- readPathoscopeData(example_data_dir, pathoreport_file_suffix)
dat <- datlist$data
ids <- rownames(dat)
tids <- unlist(lapply(ids, FUN = grepTid))
taxonLevels <- findTaxonomy(tids[1:5])
taxmat <- findTaxonMat(ids[1:5], taxonLevels)
}

