% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker.R
\name{getSignatureFromMultipleGlmnet}
\alias{getSignatureFromMultipleGlmnet}
\title{Use Lasso to do feature selection}
\usage{
getSignatureFromMultipleGlmnet(df.input, target.vec, nfolds = 10,
  logisticRegression = TRUE, nRun = 100, alpha = 1)
}
\arguments{
\item{df.input}{Row is feature, column is sample. Required}

\item{target.vec}{y vector. Required}

\item{nfolds}{glmnet CV nfolds}

\item{logisticRegression}{doing logistic regression or linear regression.}

\item{nRun}{number of glmnet runs}

\item{alpha}{same as in glmnet}
}
\description{
Use Lasso to do feature selection
}
\examples{
getSignatureFromMultipleGlmnet(df.input, target.vec)
}
