% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grepTid}
\alias{grepTid}
\title{Greps the tid from the given identifier string}
\usage{
grepTid(id)
}
\arguments{
\item{id}{Given identifier string}
}
\value{
tid string
}
\description{
Greps the tid from the given identifier string
}
\examples{
tid <- grepTid("ti|367928|org|Bifidobacterium_adolescentis_ATCC_15703")
}

