% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{PathoStat-class}
\alias{PathoStat-class}
\alias{pathostat1}
\title{PathoStat class to store PathoStat input data including phyloseq object}
\description{
Contains all currently-supported BatchQC output data classes:
}
\details{
slots:
\describe{
    \item{average_count}{a single object of class otu_tableOrNULL}
    \item{besthit_count}{a single object of class otu_tableOrNULL}
    \item{highconf_count}{a single object of class otu_tableOrNULL}
    \item{lowconf_count}{a single object of class otu_tableOrNULL}
}
}

