% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathoStat.R
\name{pathostat}
\alias{pathostat}
\title{Build PathoStat-class object from its phyloseq component.}
\usage{
pathostat(physeq1)
}
\arguments{
\item{physeq1}{phyloseq object}
}
\value{
pstat The pathostat object generated from the given phyloseq object
}
\description{
Build PathoStat-class object from its phyloseq component.
}
\examples{
rich_dense_biom  = system.file("extdata", "rich_dense_otu_table.biom", 
    package="phyloseq")
phyob <- phyloseq::import_biom(rich_dense_biom)
pstat_biom <- pathostat(phyob)
}

