% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent}
\alias{percent}
\title{Compute percentage}
\usage{
percent(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{a number or a vector}

\item{digits}{how many digit of percentage}
}
\value{
the percentage
}
\description{
Compute percentage
}
\examples{
pecent.vec <- percet(c(0.9, 0.98))
}
