% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotPCoAPlotly}
\alias{plotPCoAPlotly}
\title{Plot PCoA}
\usage{
plotPCoAPlotly(physeq.input, condition.color.vec,
  condition.color.name = "condition", condition.shape.vec,
  condition.shape.name = "condition", method = "bray",
  columnTitle = "Title", pc.a = "Axis.1", pc.b = "Axis.2")
}
\arguments{
\item{condition.color.vec}{color vector. Required}

\item{condition.color.name}{color variable name. Required}

\item{condition.shape.vec}{shape vector. Required}

\item{condition.shape.name}{shape variable name. Required}

\item{columnTitle}{Title to be displayed at top of heatmap.}

\item{df.input}{Input data object that contains the data to be plotted. Required}
}
\description{
Plot PCoA
}
\examples{
plotPCoAPlotly(physeq.input, condition.vec, condition.name = "condition", method = "bray",
columnTitle = "Title", pc.a = "Axis.1", pc.b = "Axis.2")
}
