% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readPathoscopeData}
\alias{readPathoscopeData}
\title{Reads the data from PathoScope reports and returns a list of 
final guess relative abundance and count data}
\usage{
readPathoscopeData(input_dir = ".",
    pathoreport_file_suffix = "-sam-report.tsv")
}
\arguments{
\item{input_dir}{Directory where the tsv files from PathoScope are located}

\item{pathoreport_file_suffix}{PathoScope report files suffix}
}
\value{
List of final guess relative abundance and count data
}
\description{
Reads the data from PathoScope reports and returns a list of 
final guess relative abundance and count data
}
\examples{
example_data_dir <- system.file("example/data", package = "PathoStat")
readPathoscopeData(input_dir=example_data_dir)
}

