% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathoStat.R
\name{runPathoStat}
\alias{runPathoStat}
\title{Statistical Microbiome Analysis on the pathostat input and generates a 
html report and produces interactive shiny app plots}
\usage{
runPathoStat(pstat = NULL, report_file = "PathoStat_report.html",
    report_dir = ".", report_option_binary = "111111111",
    view_report = FALSE, interactive = TRUE)
}
\arguments{
\item{pstat}{phyloseq extension pathostat object}

\item{report_file}{Output report file name}

\item{report_dir}{Output report directory path}

\item{report_option_binary}{9 bits Binary String representing the plots to 
display and hide in the report}

\item{view_report}{when TRUE, opens the report in a browser}

\item{interactive}{when TRUE, opens the interactive shinyApp}
}
\value{
outputfile The output file with all the statistical plots
}
\description{
Statistical Microbiome Analysis on the pathostat input and generates a 
html report and produces interactive shiny app plots
}
\examples{
runPathoStat(interactive = FALSE)
}

