% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{savePstat}
\alias{savePstat}
\title{Save the pathostat object to R data(.rda) file}
\usage{
savePstat(pstat, outdir = ".", outfileName = "pstat_data.rda")
}
\arguments{
\item{pstat}{pathostat object}

\item{outdir}{Output Directory of the .rda file}

\item{outfileName}{File name of the .rda file}
}
\value{
outfile .rda file
}
\description{
Save the pathostat object to R data(.rda) file
}
\examples{
data(pstat_data)
outfile <- savePstat(pstat)
}

