% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarizeTable}
\alias{summarizeTable}
\title{Summarize sample}
\usage{
summarizeTable(pstat)
}
\arguments{
\item{indata}{Input pstat}
}
\value{
A data.frame object of summary metrics.
}
\description{
Creates a table of summary metrics
}
