% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAnalysis.R
\name{Wilcox_Test_df}
\alias{Wilcox_Test_df}
\title{Mann-whitney test for a dataframe}
\usage{
Wilcox_Test_df(df, label.vec.num, pvalue.cutoff = 0.05)
}
\arguments{
\item{df}{Input data object that contains the data to be tested. Required}

\item{label.vec.num}{The target binary condition. Required}

\item{pvalue.cutoff}{choose p-value cut-off}
}
\value{
df.output object
}
\description{
Mann-whitney test for a dataframe
}
\examples{
Wilcox_Test_df(df.list[[i]], label.vec.num, pvalue.cutoff)
}
